hiding LibraryMembership serialization and using more of LibraryMembershipView
